# backend/src/main/resources/api-docs.yaml
openapi: 3.0.3
info:
  title: MyTrip API
  description: API documentation for MyTrip travel booking platform
  version: 1.0.0
  contact:
    name: MyTrip Support
    email: support@mytrip.com
servers:
  - url: http://localhost:8080/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /reviews:
    get:
      summary: Get all reviews
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
    post:
      summary: Create a new review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Bad request (invalid data)
        '401':
          description: Unauthorized
...
  /reviews/hotel/{hotelId}:
    get:
      summary: Get reviews for a hotel
      parameters:
        - in: path
          name: hotelId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of reviews for the hotel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          description: Hotel not found
  /reviews/flight/{flightId}:
    get:
      summary: Get reviews for a flight
      parameters:
        - in: path
          name: flightId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of reviews for the flight
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          description: Flight not found
  /reviews/{id}/flag:
    put:
      summary: Flag a review as inappropriate
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Review flagged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          description: Unauthorized
        '404':
          description: Review not found
  /reviews/{id}/moderate:
    put:
      summary: Moderate a flagged review
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                moderatorNotes:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Review moderated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (admin only)
        '404':
          description: Review not found
...